% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{colVars}
\alias{colVars}
\title{Column Variances}
\usage{
colVars(x, na.rm = FALSE, dims = 1, unbiased = TRUE, SumSquares = FALSE,
  twopass = FALSE)
}
\arguments{
\item{x}{matrix}

\item{twopass}{}
}
\description{
Calculate variance for each column of given matrix
}
