% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.r
\name{create_beta_equations}
\alias{create_beta_equations}
\title{GEE Beta Equations}
\usage{
create_beta_equations(betas, rho, dat)
}
\arguments{
\item{betas}{vector of coefficients}

\item{rho}{overdispersion parameter}

\item{dat}{matrix of multinomial clusters}
}
\value{
estimating equations for betas
}
\description{
Creates GEE Equations for beta coefficients.
}
