% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.r
\name{create_rho_equations}
\alias{create_rho_equations}
\title{GEE Rho Equations}
\usage{
create_rho_equations(rho, betas, dat)
}
\arguments{
\item{rho}{overdispersion parameter}

\item{betas}{vector of coefficients}

\item{dat}{matrix of multinomial clusters}
}
\value{
estimating equations for rho
}
\description{
Creates GEE Equations for rho (overdispersion parameter).
}
