% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.r
\name{CreateBetaEquations}
\alias{CreateBetaEquations}
\title{GEE Beta Equations}
\usage{
CreateBetaEquations(betas, rho, data)
}
\arguments{
\item{betas}{vector of coefficients}

\item{rho}{overdispersion parameter}

\item{data}{matrix of multinomial clusters}
}
\value{
estimating equations for betas
}
\description{
Creates GEE Equations for beta coefficients.
}
