% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations.r
\name{CreateRhoEquations}
\alias{CreateRhoEquations}
\title{GEE Rho Equations}
\usage{
CreateRhoEquations(rho, betas, data)
}
\arguments{
\item{rho}{overdispersion parameter}

\item{betas}{vector of coefficients}

\item{data}{matrix of multinomial clusters}
}
\value{
estimating equations for rho
}
\description{
Creates GEE Equations for rho (overdispersion parameter).
}
