% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.r
\name{get_var_ident}
\alias{get_var_ident}
\title{Get GEE Identity Co-Variance Matrix}
\usage{
get_var_ident(p.vec, rho, dat)
}
\arguments{
\item{p.vec}{vector of probabilities}

\item{rho}{overdispersion parameter}

\item{dat}{matrix of multinomial clusters}
}
\value{
matrix of covariance
}
\description{
Calculates GEE identity variance co-variance matrix.
}
