% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.r
\name{GetIdentityVarCov}
\alias{GetIdentityVarCov}
\title{Get GEE Identity Co-Variance Matrix}
\usage{
GetIdentityVarCov(p.vec, rho, dat)
}
\arguments{
\item{p.vec}{vector of probabilities}

\item{rho}{overdispersion parameter}

\item{dat}{matrix of multinomial clusters}
}
\value{
matrix of covariance
}
\description{
Calculates GEE identity variance co-variance matrix.
}
