% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom.r
\name{gom}
\alias{gom}
\title{GOM}
\usage{
gom(dat, corstr = "independence", method = "GEE", ...)
}
\arguments{
\item{dat}{matrix of multinomial clusters}

\item{corstr}{one of: independence, exchangeable}

\item{method}{one of: GEE, DM, RCM}

\item{...}{parameters to be passed on to \code{IterativeFunction}}
}
\value{
a GOM object
}
\description{
Creates an General Overdispersion Model for the given data
}
\examples{
dat <- matrix(c(1, 1, 1, 3, 0, 0, 2, 1, 0), 3, byrow = TRUE)
gom(dat)
}
