% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.r
\name{IterativeFunction}
\alias{IterativeFunction}
\title{Iterative GEE Function}
\usage{
IterativeFunction(num.iter, thresholds, data, RHO = FALSE)
}
\arguments{
\item{num.iter}{max number of iterations}

\item{thresholds}{vector of thresholds for betas and rho}

\item{data}{matrix of multinomial clusters}

\item{RHO}{if FALSE, rho will not be estimated}
}
\value{
...
}
\description{
Iteratively finds coefficients and overdispersion parameter using
estimating equations.
}
