% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmo.r
\name{print.gmo.ident}
\alias{print.gmo.ident}
\title{Print GMO identity}
\usage{
\method{print}{gmo.ident}(x, ...)
}
\arguments{
\item{x}{gom identity object}

\item{...}{additional params to \code{print(x, ...)}}
}
\value{
None
}
\description{
Prints GMO identity object
}
