% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gom.r
\name{print.gom.gee}
\alias{print.gom.gee}
\title{Print GOM GEE}
\usage{
\method{print}{gom.gee}(gom.gee)
}
\arguments{
\item{gom.gee}{om gee object}
}
\value{
None
}
\description{
Prints OM GEE object
}
\examples{
dat <- matrix(c(1, 1, 1, 3, 0, 0, 2, 1, 0), 3, byrow = TRUE)
res <- gom(dat, method = "GEE")
print(res)
}
