% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{tr}
\alias{tr}
\title{Compute trace of a square matrix, that is, the sum
of its diagonal elements.}
\usage{
tr(m)
}
\arguments{
\item{m}{a square matrix}
}
\value{
trace of input matrix
}
\description{
Compute trace of a square matrix, that is, the sum
of its diagonal elements.
}
