% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.r
\name{GetIdentityVariance}
\alias{GetIdentityVariance}
\title{Get GEE Identity Variance}
\usage{
GetIdentityVariance(p.vec, rho, data)
}
\arguments{
\item{p.vec}{vector of probabilities}

\item{rho}{overdispersion parameter}

\item{data}{matrix of multinomial clusters}
}
\value{
standard error
}
\description{
Calculates GEE identity variance.
}
