% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmo.r
\name{gmo}
\alias{gmo}
\title{GMO}
\usage{
gmo(dat, corstr = "independence", link = "glogit", ...)
}
\arguments{
\item{dat}{matrix of multinomial clusters}

\item{corstr}{one of: independence, exchangeable}

\item{link}{one of: glogit, identity}

\item{...}{parameters to be passed on to \code{iterative_function}}
}
\value{
a GMO object
}
\description{
Creates an General Multinomial Overdispersion Model for the given data
}
\examples{
dat <- matrix(c(1, 1, 1, 3, 0, 0, 2, 1, 0), 3, byrow = TRUE)
gom(dat)
}
