% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.r
\name{IterativeFunction}
\alias{IterativeFunction}
\title{Iterative GEE Function}
\usage{
IterativeFunction(dat, RHO = FALSE, num.iter = 10, thresholds = NULL)
}
\arguments{
\item{dat}{matrix of multinomial clusters}

\item{RHO}{if FALSE, rho will not be estimated}

\item{num.iter}{max number of iterations}

\item{thresholds}{vector of thresholds for betas and rho}
}
\value{
list containing estimates (probabilities and rho) and number of iterations
}
\description{
Iteratively finds coefficients and overdispersion parameter using
estimating equations.
}
