% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmo.r
\name{print.gmo.glogit}
\alias{print.gmo.glogit}
\title{Print GMO glogit}
\usage{
\method{print}{gmo.glogit}(x, ...)
}
\arguments{
\item{x}{gom glogit object}

\item{...}{additional params to \code{print(x, ...)}}
}
\value{
None
}
\description{
Prints GMO glogit object.
}
